Amended Plugin.pm to simplify install
I have re-worked a lot of the perl scripts to be compliant with squeezebox install standards

This means that the binaries should be distributed in the install zip and "live" on  ..\Cache\InstalledPlugins\Plugins\InguzEQ\Bin
Also preferences have been moved so that they now live in ..\prefs\InguzEQ

I have amended the Inguz binaries so that they are now looking in the correct directories. I have maintained a degree of backward compatibility as it will default to the original folders if Cache or cache is not found in the file path.

There is some additional code to detect binary locations and if necessary to copy them into the Bin folder and to set the permissions of the files to be executable.

I have done this because multiple versions of the binaries are installed for each supported OS, but I wanted to use a single name for the binary so that the app config still works.
I could have copied multiple app configs, but they may well have got out of step.
On linux all the files and folders will need to be owned by squeezeboxserver otherwise copying and changing mode do not work. The installer should handle this.